refactor: 중복 컴포넌트 제거 및 폴더 경로 이동
